+++ resources/views/project/project_manager/pages/project/module/index.blade.php
@@ -19,14 +19,16 @@
                 <div class="card-header">
                     <h4>Project Modules <span class="badge badge-secondary">{{ $projectModules->count() }}</span></h4>
                     <div class="card-header-action">
-                        <button data-action="{{ route('project_manager.projects.module.store', $project) }}" class="btn btn-primary btn-round ml-auto btn-add text-white" data-toggle="modal" data-target="#modal">
-                            <i class="fa fa-plus"></i>
-                            Add Module
-                        </button>
-                        <button data-action="{{ route('project_manager.projects.module.special.store', $project) }}" class="btn btn-outline-primary btn-round ml-auto btn-special-module" data-toggle="modal" data-target="#modalSpecial">
-                            <i class="fa fa-plus"></i>
-                            Add Special Module
-                        </button>
+                        @if($project->user_id == Auth::user()->id)
+                            <button data-action="{{ route('project_manager.projects.module.store', $project) }}" class="btn btn-primary btn-round ml-auto btn-add text-white" data-toggle="modal" data-target="#modal">
+                                <i class="fa fa-plus"></i>
+                                Add Module
+                            </button>
+                            <button data-action="{{ route('project_manager.projects.module.special.store', $project) }}" class="btn btn-outline-primary btn-round ml-auto btn-special-module" data-toggle="modal" data-target="#modalSpecial">
+                                <i class="fa fa-plus"></i>
+                                Add Special Module
+                            </button>
+                        @endif
                     </div>
                 </div>
                 <div class="card-body">
@@ -37,8 +39,8 @@
                                     <th>Name</th>
                                     <th>Status</th>
                                     <th>Assignee</th>
-                                    <th>Start Date</th>
-                                    <th>End Date</th>
+                                    <th>Start Date Plan</th>
+                                    <th>End Date Plan</th>
                                     <th>Action</th>
                                 </tr>
                             </thead>
@@ -97,27 +99,30 @@
                                             </td>
                                             <td>
                                                 <a href="{{ route('project_manager.projects.module.show', $projectModule) }}" class="btn btn-light" data-toggle="tooltip" title="View"><i class="fas fa-eye"></i></a>
-                                                <a data-action="{{ route('project_manager.projects.module.member.store', $projectModule) }}" title="Add Member" href="#" class="btn btn-primary btn-add-member" data-toggle="modal" data-target="#modalMember" title="Add Member"><i class="fa fa-user-plus"></i></a>
 
-                                                @if($projectModule->moduleable_type == $projectModule->moduleType['special_module'])
-                                                    <a data-detail="{{ route('project_manager.projects.module.edit', $projectModule) }}" data-action="{{ route('project_manager.projects.module.special.update', $projectModule) }}" href="#" class="btn btn-primary btn-edit-special" data-toggle="modal" data-target="#modalSpecial" title="Edit"><i class="fa fa-pencil-alt"></i></a>
+                                                @if($project->user_id == Auth::user()->id)
+                                                    <a data-action="{{ route('project_manager.projects.module.member.store', $projectModule) }}" title="Add Member" href="#" class="btn btn-primary btn-add-member" data-toggle="modal" data-target="#modalMember" title="Add Member"><i class="fa fa-user-plus"></i></a>
 
-                                                    <a href="#" onclick="deleteConfirm('del{{ $projectModule->id }}')" class="btn btn-danger btn-action" data-toggle="tooltip" title="Delete">
-                                                        <i class="fa fa-trash"></i>
-                                                    </a>
-                                                    <form id="del{{ $projectModule->id }}" action="{{ route('project_manager.projects.module.special.destroy', $projectModule) }}" method="POST">
-                                                        @method('delete')
-                                                        @csrf
-                                                    </form>
-                                                @else
-                                                    <a data-detail="{{ route('project_manager.projects.module.edit', $projectModule) }}" data-action="{{ route('project_manager.projects.module.update', $projectModule) }}" href="#" class="btn btn-primary btn-edit" data-toggle="modal" data-target="#modal" title="Edit"><i class="fa fa-pencil-alt"></i></a>
-                                                    <a href="#" onclick="deleteConfirm('del{{ $projectModule->id }}')" class="btn btn-danger btn-action" data-toggle="tooltip" title="Delete">
-                                                        <i class="fa fa-trash"></i>
-                                                    </a>
-                                                    <form id="del{{ $projectModule->id }}" action="{{ route('project_manager.projects.module.destroy', $projectModule) }}" method="POST">
-                                                        @method('delete')
-                                                        @csrf
-                                                    </form>
+                                                    @if($projectModule->moduleable_type == $projectModule->moduleType['special_module'])
+                                                        <a data-detail="{{ route('project_manager.projects.module.edit', $projectModule) }}" data-action="{{ route('project_manager.projects.module.special.update', $projectModule) }}" href="#" class="btn btn-primary btn-edit-special" data-toggle="modal" data-target="#modalSpecial" title="Edit"><i class="fa fa-pencil-alt"></i></a>
+
+                                                        <a href="#" onclick="deleteConfirm('del{{ $projectModule->id }}')" class="btn btn-danger btn-action" data-toggle="tooltip" title="Delete">
+                                                            <i class="fa fa-trash"></i>
+                                                        </a>
+                                                        <form id="del{{ $projectModule->id }}" action="{{ route('project_manager.projects.module.special.destroy', $projectModule) }}" method="POST">
+                                                            @method('delete')
+                                                            @csrf
+                                                        </form>
+                                                    @else
+                                                        <a data-detail="{{ route('project_manager.projects.module.edit', $projectModule) }}" data-action="{{ route('project_manager.projects.module.update', $projectModule) }}" href="#" class="btn btn-primary btn-edit" data-toggle="modal" data-target="#modal" title="Edit"><i class="fa fa-pencil-alt"></i></a>
+                                                        <a href="#" onclick="deleteConfirm('del{{ $projectModule->id }}')" class="btn btn-danger btn-action" data-toggle="tooltip" title="Delete">
+                                                            <i class="fa fa-trash"></i>
+                                                        </a>
+                                                        <form id="del{{ $projectModule->id }}" action="{{ route('project_manager.projects.module.destroy', $projectModule) }}" method="POST">
+                                                            @method('delete')
+                                                            @csrf
+                                                        </form>
+                                                    @endif
                                                 @endif
                                             </td>
                                         </tr>
@@ -156,21 +161,25 @@
                                 <small class="text-danger">{{ $message }}</small>
                             @enderror
                         </div>
-                        <div class="form-row">
-                            <div class="form-group col-md-6 col-lg-6">
-                                <label>Start Date</label>
-                                <input value="{{ old('start_date') }}" id="start" type="text" name="start_date" class="form-control datetimepicker">
-                                @error('start_date')
-                                    <small class="text-danger">{{ $message }}</small>
-                                @enderror
+                        <div class="form-group">
+                            <label>Start - End Date Plan</label>
+                            <div class="input-group">
+                                <div class="input-group-prepend">
+                                    <div class="input-group-text">
+                                        <i class="fas fa-calendar"></i>
+                                    </div>
+                                </div>
+                                <input
+                                    type="text"
+                                    name="start_end_date"
+                                    id="startEndDate"
+                                    class="form-control daterange-cus"
+                                    value=""
+                                />
                             </div>
-                            <div class="form-group col-md-6 col-lg-6">
-                                <label>End Date</label>
-                                <input value="{{ old('end_date') }}" id="end" type="text" name="end_date" class="form-control datetimepicker">
-                                @error('end_date')
-                                    <small class="text-danger">{{ $message }}</small>
-                                @enderror
-                            </div>
+                            @error('start_end_date')
+                                <small class="text-danger">{{ $message }}</small>
+                            @enderror
                         </div>
                     </div>
                     <div class="modal-footer bg-whitesmoke br">
@@ -209,21 +218,25 @@
                                 <small class="text-danger">{{ $message }}</small>
                             @enderror
                         </div>
-                        <div class="form-row">
-                            <div class="form-group col-md-6 col-lg-6">
-                                <label>Start Date</label>
-                                <input value="{{ old('start_date') }}" type="text" id="startSpecial" name="start_date" class="form-control datetimepicker">
-                                @error('start_date')
-                                    <small class="text-danger">{{ $message }}</small>
-                                @enderror
+                        <div class="form-group">
+                            <label>Start - End Date Plan</label>
+                            <div class="input-group">
+                                <div class="input-group-prepend">
+                                    <div class="input-group-text">
+                                        <i class="fas fa-calendar"></i>
+                                    </div>
+                                </div>
+                                <input
+                                    type="text"
+                                    name="start_end_date"
+                                    id="startEndDateSpecial"
+                                    class="form-control daterange-cus"
+                                    value=""
+                                />
                             </div>
-                            <div class="form-group col-md-6 col-lg-6">
-                                <label>End Date</label>
-                                <input value="{{ old('end_date') }}" type="text" id="endSpecial" name="end_date" class="form-control datetimepicker">
-                                @error('end_date')
-                                    <small class="text-danger">{{ $message }}</small>
-                                @enderror
-                            </div>
+                            @error('start_end_date')
+                                <small class="text-danger">{{ $message }}</small>
+                            @enderror
                         </div>
                     </div>
                     <div class="modal-footer bg-whitesmoke br">
@@ -311,6 +324,25 @@
     <script src="{{ asset('templates/stisla/node_modules/sweetalert/dist/sweetalert.min.js') }}"></script>
     <script src="{{ asset('templates/stisla/node_modules/bootstrap-daterangepicker/daterangepicker.js') }}"></script>
     <script src="{{ asset('templates/stisla/node_modules/bootstrap-timepicker/js/bootstrap-timepicker.min.js') }}"></script>
+    <script>
+        $(function() {
+            $('#startEndDate').daterangepicker({
+                timePicker: true,
+                locale: {
+                    format: 'YYYY-MM-DD HH:mm'
+                }
+            });
+        });
+
+        $(function() {
+            $('#startEndDateSpecial').daterangepicker({
+                timePicker: true,
+                locale: {
+                    format: 'YYYY-MM-DD HH:mm'
+                }
+            });
+        });
+    </script>
     @if (Session::has('success'))
         <script>
             swal("Success!", "{{ Session::get('success') }}", "success").then(function(){
